Renamed modules to better describe their purposes
Added version information in the submodules
